+++ canvas_cli/tests.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 from collections.abc import Callable, Generator
+from contextlib import chdir
 from datetime import datetime
 from pathlib import Path
 from typing import Any, cast
@@ -71,9 +72,14 @@ def create_or_update_config_auth_file_for_testing(plugin_name: str) -> Generator
 
 
 @pytest.fixture(autouse=True, scope="session")
-def write_plugin(cli_runner: CliRunner, plugin_name: str) -> Generator[Any, Any, Any]:
+def write_plugin(
+    cli_runner: CliRunner,
+    integration_tests_plugins_dir: Path,
+    plugin_name: str,
+) -> Generator[Any, Any, Any]:
     """Writes a plugin to the file system."""
-    cli_runner.invoke(app, "init", input=plugin_name)
+    with chdir(integration_tests_plugins_dir):
+        cli_runner.invoke(app, "init", input=plugin_name)
 
     protocol_code = """
 from canvas_sdk.events import EventType
@@ -88,14 +94,15 @@ class Protocol(BaseProtocol):
         log.info(self.NARRATIVE_STRING)
         return []
 """
+    plugin_dir = integration_tests_plugins_dir / plugin_name
 
-    with open(f"./{plugin_name}/protocols/my_protocol.py", "w") as protocol:
+    with open(plugin_dir / "protocols" / "my_protocol.py", "w") as protocol:
         protocol.write(protocol_code)
 
     yield
 
-    if Path(f"./{plugin_name}").exists():
-        shutil.rmtree(Path(f"./{plugin_name}"))
+    if plugin_dir.exists():
+        shutil.rmtree(plugin_dir)
 
 
 def list_empty_plugins(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
@@ -112,7 +119,7 @@ def install_new_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]
             f"Plugin {plugin_name} has a valid CANVAS_MANIFEST.json file",
             "Installing plugin:",
             "Posting",
-            f"Plugin {plugin_name} successfully installed!",
+            f"Plugin {plugin_name} uploaded! Check logs for more details.",
         ],
         [],
     )
@@ -123,8 +130,41 @@ def list_newly_installed_plugin(plugin_name: str) -> tuple[str, int, list[str], 
     return ("list", 0, [f"{plugin_name}@0.0.1	enabled"], [])
 
 
+def reinstall_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
+    """Step 4 - make a change and reinstall the plugin."""
+    protocol_code = """
+from canvas_sdk.events import EventType
+from canvas_sdk.protocols import BaseProtocol
+from logger import log
+
+class Protocol(BaseProtocol):
+    RESPONDS_TO = EventType.Name(EventType.ASSESS_COMMAND__CONDITION_SELECTED)
+    NARRATIVE_STRING = "EDITED PROTOCOL: I was inserted from my plugin's protocol."
+
+    def compute(self):
+        log.info(self.NARRATIVE_STRING)
+        return []
+"""
+
+    with open(f"./{plugin_name}/protocols/my_protocol.py", "w") as protocol:
+        protocol.write(protocol_code)
+
+    return (
+        f"install {plugin_name}",
+        0,
+        [
+            f"Plugin {plugin_name} has a valid CANVAS_MANIFEST.json file",
+            "Installing plugin:",
+            "Posting",
+            f"Plugin {plugin_name} already exists, updating instead...",
+            "New plugin version uploaded! Check logs for more details.",
+        ],
+        [],
+    )
+
+
 def disable_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
-    """Step 4 - disable plugin."""
+    """Step 5 - disable plugin."""
     return (
         f"disable {plugin_name}",
         0,
@@ -134,22 +174,25 @@ def disable_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
 
 
 def list_disabled_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
-    """Step 5 - list disabled plugin."""
+    """Step 6 - list disabled plugin."""
     return ("list", 0, [f"{plugin_name}@0.0.1	disabled"], [])
 
 
 def enable_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
-    """Step 6 - enable the disabled plugin."""
+    """Step 7 - enable the disabled plugin."""
     return (
         f"enable {plugin_name}",
         0,
-        [f"Enabling {plugin_name} using ", f"Plugin {plugin_name} successfully enabled!"],
+        [
+            f"Enabling {plugin_name} using ",
+            f"Plugin {plugin_name} successfully enabled!",
+        ],
         [],
     )
 
 
 def uninstall_enabled_plugin(plugin_name: str) -> tuple[str, int, list[str], list[str]]:
-    """Step 7 - try to uninstall the enabled plugin."""
+    """Step 8 - try to uninstall the enabled plugin."""
     return (
         f"uninstall {plugin_name}",
         1,
@@ -183,6 +226,7 @@ def uninstall_disabled_plugin(plugin_name: str) -> tuple[str, int, list[str], li
         (list_empty_plugins),
         (install_new_plugin),
         (list_newly_installed_plugin),
+        (reinstall_plugin),
         (disable_plugin),
         (list_disabled_plugin),
         (enable_plugin),
@@ -195,21 +239,22 @@ def uninstall_disabled_plugin(plugin_name: str) -> tuple[str, int, list[str], li
 def test_canvas_list_install_disable_enable_uninstall(
     mock_get_token: MagicMock,
     mock_set_token: MagicMock,
+    integration_tests_plugins_dir: Path,
     plugin_name: str,
     create_or_update_config_auth_file_for_testing: None,
     step: Callable,
     cli_runner: CliRunner,
 ) -> None:
-    """Tests that the Canvas CLI can list, install, disable, enable, and uninstall a plugin."""
+    """Tests that the Canvas CLI can list, install, reinstall, disable, enable, and uninstall a plugin."""
     mock_get_token.return_value = None
     mock_set_token.return_value = None
 
-    (command, expected_exit_code, expected_outputs, expected_no_outputs) = step(plugin_name)
+    with chdir(integration_tests_plugins_dir):
+        (command, expected_exit_code, expected_outputs, expected_no_outputs) = step(plugin_name)
+        result = cli_runner.invoke(app, command)
 
-    result = cli_runner.invoke(app, command)
-
-    assert result.exit_code == expected_exit_code
-    for expected_output in expected_outputs:
-        assert expected_output in result.stdout
-    for expected_no_output in expected_no_outputs:
-        assert expected_no_output not in result.stdout
+        assert result.exit_code == expected_exit_code
+        for expected_output in expected_outputs:
+            assert expected_output in result.stdout
+        for expected_no_output in expected_no_outputs:
+            assert expected_no_output not in result.stdout