+++ canvas_sdk/v1/data/__init__.py
@@ -13,6 +13,7 @@ from .charge_description_master import ChargeDescriptionMaster
 from .claim import Claim, ClaimCoverage, ClaimPatient, ClaimQueue, InstallmentPlan
 from .claim_line_item import ClaimLineItem
 from .command import Command
+from .compound_medication import CompoundMedication
 from .condition import Condition, ConditionCoding
 from .coverage import Coverage, Transactor, TransactorAddress, TransactorPhone
 from .detected_issue import DetectedIssue, DetectedIssueEvidence
@@ -83,7 +84,9 @@ from .questionnaire import (
     ResponseOptionSet,
 )
 from .reason_for_visit import ReasonForVisitSettingCoding
-from .staff import Staff, StaffAddress, StaffContactPoint, StaffPhoto
+from .referral import Referral, ReferralReport
+from .service_provider import ServiceProvider
+from .staff import Staff, StaffAddress, StaffContactPoint, StaffPhoto, StaffRole
 from .task import Task, TaskComment, TaskLabel, TaskTaskLabel
 from .team import Team, TeamContactPoint
 from .user import CanvasUser
@@ -112,6 +115,7 @@ __all__ = __exports__ = (
     "ClaimPatient",
     "ClaimQueue",
     "Command",
+    "CompoundMedication",
     "Condition",
     "ConditionCoding",
     "Coverage",
@@ -175,11 +179,15 @@ __all__ = __exports__ = (
     "Questionnaire",
     "QuestionnaireQuestionMap",
     "ReasonForVisitSettingCoding",
+    "Referral",
+    "ReferralReport",
     "ResponseOption",
     "ResponseOptionSet",
+    "ServiceProvider",
     "Staff",
     "StaffAddress",
     "StaffPhoto",
+    "StaffRole",
     "StaffContactPoint",
     "Task",
     "TaskComment",