+++ plugin_runner/sandbox.py
@@ -4,6 +4,7 @@ import ast
 import builtins
 import importlib
 import json
+import operator
 import sys
 import types
 from _ast import AnnAssign
@@ -196,6 +197,8 @@ THIRD_PARTY_MODULES = {
         "BigIntegerField",
         "Case",
         "CharField",
+        "Count",
+        "F",
         "IntegerField",
         "Model",  # remove when hyperscribe no longer needs it
         "Q",
@@ -216,8 +219,17 @@ THIRD_PARTY_MODULES = {
     "jwt": {
         "decode",
         "encode",
+        "PyJWKClient",
+        "ExpiredSignatureError",
+        "InvalidTokenError",
     },
     "pydantic": {
+        "BaseModel",
+        "ConfigDict",
+        "conint",
+        "constr",
+        "Field",
+        "RootModel",
         "ValidationError",
     },
     "rapidfuzz": {
@@ -255,6 +267,34 @@ def _is_known_module(name: str) -> bool:
 def _unrestricted(_ob: Any, *args: Any, **kwargs: Any) -> Any:
     """Return the given object, unmodified."""
     return _ob
+
+
+def _inplacevar(op: Any, val: Any, expr: Any) -> Any:
+    """
+    Apply the specified operation to the value and expression.
+
+    NOTE: Does not yet support += concatenation for sequences.
+    """
+    ops = {
+        "-=": operator.isub,
+        "@=": operator.imatmul,
+        "**=": operator.ipow,
+        "*=": operator.imul,
+        "//=": operator.ifloordiv,
+        "/=": operator.itruediv,
+        "&=": operator.iand,
+        "%=": operator.imod,
+        "^=": operator.ixor,
+        "+=": operator.iadd,
+        "<<=": operator.ilshift,
+        ">>=": operator.irshift,
+        "|=": operator.ior,
+    }
+
+    if op not in ops:
+        raise ValueError(f"Invalid inplace operation: {op}")
+
+    return ops[op](val, expr)
 
 
 def _apply(_ob: Any, *args: Any, **kwargs: Any) -> Any:
@@ -349,6 +389,10 @@ class Sandbox:
 
         def visit_AnnAssign(self, node: AnnAssign) -> AnnAssign:
             """Allow type annotations."""
+            return self.node_contents_visit(node)
+
+        def visit_Constant(self, node: ast.Constant) -> ast.Constant:
+            """Allow constants."""
             return self.node_contents_visit(node)
 
         def visit_Match(self, node: ast.Match) -> ast.Match:
@@ -603,6 +647,7 @@ class Sandbox:
                 "dict": builtins.dict,
                 "enumerate": builtins.enumerate,
                 "filter": builtins.filter,
+                "getattr": self._safe_getattr,
                 "hasattr": builtins.hasattr,
                 "iter": builtins.iter,
                 "list": builtins.list,
@@ -623,7 +668,7 @@ class Sandbox:
             "_getattr_": self._safe_getattr,
             "_getitem_": self._safe_getitem,
             "_getiter_": _unrestricted,
-            "_inplacevar_": _unrestricted,
+            "_inplacevar_": _inplacevar,
             "_iter_unpack_sequence_": guarded_iter_unpack_sequence,
             "_print_": PrintCollector,
             "_unpack_sequence_": guarded_unpack_sequence,